addressing PR comments, plus some refactoring to fix the tests (mostly), still not green
