+++ lib/model/classycat.py
@@ -1,24 +1,12 @@
-from typing import Union, Dict, Any, Optional, List
-from pydantic import BaseModel
+from typing import Union, Dict, Any
 from lib.logger import logger
 from lib.model.model import Model
 from lib.schemas import Message
 from lib.model.classycat_classify import Model as ClassifyModel
 from lib.model.classycat_schema_create import Model as ClassyCatSchemaCreateModel
 from lib.model.classycat_schema_lookup import Model as ClassyCatSchemaLookupModel
+from lib.model.classycat_response import ClassyCatSchemaResponse, ClassyCatBatchClassificationResponse
 from lib.base_exception import PrestoBaseException
-
-
-class ClassyCatResponse(BaseModel):
-    responseMessage: Optional[str] = None
-
-
-class ClassyCatBatchClassificationResponse(ClassyCatResponse):
-    classification_results: Optional[List[dict]] = []
-
-
-class ClassyCatSchemaResponse(ClassyCatResponse):
-    schema_id: Optional[str] = None
 
 
 class Model(Model):
