+++ lib/model/classycat_classify.py
@@ -8,7 +8,7 @@ from anthropic import Anthropic
 from lib.logger import logger
 from lib.model.model import Model
 from lib.schemas import Message
-from lib.model.classycat import ClassyCatBatchClassificationResponse
+from lib.model.classycat_response import ClassyCatBatchClassificationResponse
 from lib.s3 import load_file_from_s3, file_exists_in_s3, upload_file_to_s3
 from lib.base_exception import PrestoBaseException
