+++ lib/model/classycat_response.py
@@ -1,0 +1,14 @@
+from typing import Optional, List
+from pydantic import BaseModel
+
+
+class ClassyCatResponse(BaseModel):
+    responseMessage: Optional[str] = None
+
+
+class ClassyCatBatchClassificationResponse(ClassyCatResponse):
+    classification_results: Optional[List[dict]] = []
+
+
+class ClassyCatSchemaResponse(ClassyCatResponse):
+    schema_id: Optional[str] = None
