+++ lib/model/classycat_schema_create.py
@@ -6,7 +6,7 @@ from lib.s3 import upload_file_to_s3, file_exists_in_s3
 from lib.logger import logger
 from lib.model.model import Model
 from lib.schemas import Message
-from lib.model.classycat import ClassyCatSchemaResponse
+from lib.model.classycat_response import ClassyCatSchemaResponse
 from lib.base_exception import PrestoBaseException
